+++ frontend/src/common/mock/handlers.ts
@@ -3,6 +3,7 @@ import { editProfileHandlers } from '../../pages/editProfile/mock/handler';
 import { loginHandler } from '../../pages/login/mock/handler';
 import { mentoringCreateHandler } from '../../pages/mentoringCreate/mock/handlers';
 import { mentoringUpdateHandler } from '../../pages/mentoringUpdate/mock/handlers';
+import { signupHandler } from '../../pages/signup/mock/signup/handlers';
 import { validateIdHandler } from '../../pages/signup/mock/validateId/handler';
 
 import { authCodeHandler } from './authCode/authCode';
@@ -10,7 +11,6 @@ import { authCodeVerifyHandler } from './authCodeVerify/authCodeVerify';
 import { commonHandler } from './common/handlers';
 import { mentoringHandler } from './mentoring/handlers';
 import { participatedMentoringHandler } from './participatedMentoring/handler';
-import { signupHandler } from './signup/signup';
 
 export const handlers = [
   ...mentoringHandler,
